+++ myWeb/App_Control/budget_money/budget_money_control.aspx.cs
@@ -38,7 +38,7 @@ namespace myWeb.App_Control.budget_money
         
         public static string getNumber(object pNumber)
         {
-            string strNumber = String.Format("{0:#,##0.00}", float.Parse(pNumber.ToString()));
+            string strNumber = String.Format("{0:#,##0.00}", decimal.Parse(pNumber.ToString()));
             return strNumber;
         }
 
@@ -139,7 +139,7 @@ namespace myWeb.App_Control.budget_money
                 //        base.SetLabel(c, "แผนงาน", "งานย่อย");
                 //        base.SetLabel(c, "กิจกรรม", "งานรอง");
                 //        base.SetLabel(c, "แผนงบประมาณ", "แผนงาน");
-                //        base.SetLabel(c, "ผลผลิต", "งานหลัก");
+                //        base.SetLabel(c, "รายการ", "งานหลัก");
                 //    }
                 //}
